+++ src/bridge/rsa/rsa_test.cc
@@ -24,7 +24,7 @@
 
 #include "src/bridge/ex_data_util/ex_data_util.h"
 #include "src/bridge/memory_util/openssl_structs.h"
-#include "src/bridge/rsa/rsa_method.h"
+#include "src/bridge/rsa/rsa.h"
 #include "src/testing_util/mock_rsa_key.h"
 #include "src/testing_util/openssl_assertions.h"
 #include "src/testing_util/test_matchers.h"
